fix initialization problems with script
This is probably not the final factoring of script_init. I'll look again
tomorrow.
