+++ tests/__init__.py
@@ -1,2 +1,2 @@
-from brave.core.util.script_init import script_init
+from brave.core.scripts import script_init
 script_init("config:test.ini")
